feat: use tshy
