+++ package.json
@@ -1,14 +1,15 @@
 {
   "name": "mcp-proxy",
   "version": "1.0.0",
-  "main": "dist/index.js",
+  "main": "./dist/commonjs/index.js",
   "scripts": {
-    "build": "tsup",
-    "test": "vitest run && tsc && prettier --check . && eslint .",
+    "build": "tshy && tshy-after",
+    "clean": "tsc -b --clean",
+    "test": "vitest run && prettier --check . && eslint .",
     "format": "prettier --write . && eslint --fix ."
   },
   "bin": {
-    "mcp-proxy": "dist/bin/mcp-proxy.js"
+    "mcp-proxy": "./dist/commonjs/bin/mcp-proxy.js"
   },
   "keywords": [
     "MCP",
@@ -19,8 +20,8 @@
   "author": "Frank Fiegel <frank@glama.ai>",
   "license": "MIT",
   "description": "A TypeScript SSE proxy for MCP servers that use stdio transport.",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "module": "./dist/esm/index.js",
+  "types": "./dist/commonjs/index.d.ts",
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.10.1",
     "eventsource": "^3.0.6",
@@ -41,6 +42,9 @@
       "@semantic-release/github"
     ]
   },
+  "files": [
+    "dist"
+  ],
   "devDependencies": {
     "@eslint/js": "^9.25.1",
     "@sebbo2002/semantic-release-jsr": "^2.0.5",
@@ -56,7 +60,8 @@
     "jiti": "^2.4.2",
     "prettier": "^3.5.3",
     "semantic-release": "^24.2.3",
-    "tsup": "^8.4.0",
+    "tshy": "^1.1.1",
+    "tshy-after": "^1.0.0",
     "tsx": "^4.19.3",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.31.1",
@@ -74,5 +79,26 @@
     "splitting": true,
     "sourcemap": true,
     "clean": true
+  },
+  "tshy": {
+    "exports": {
+      "./package.json": "./package.json",
+      ".": "./src/index.ts"
+    }
+  },
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "import": {
+        "source": "./src/index.ts",
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.js"
+      },
+      "require": {
+        "source": "./src/index.ts",
+        "types": "./dist/commonjs/index.d.ts",
+        "default": "./dist/commonjs/index.js"
+      }
+    }
   }
 }
