--- src/simple-stdio-proxy-server.ts
@@ -1,3 +1,0 @@
-import { startStdioServer } from "./startStdioServer.js";
-
-await startStdioServer(JSON.parse(process.argv[2]));
