+++ test/simple-stdio-proxy-server.ts
@@ -1,0 +1,3 @@
+import { startStdioServer } from "../src/startStdioServer.js";
+
+await startStdioServer(JSON.parse(process.argv[2]));
