+++ test/simple-stdio-server.ts
