+++ test/startSSEServer.test.ts
@@ -1,0 +1,130 @@
+import { Client } from "@modelcontextprotocol/sdk/client/index.js";
+import { SSEClientTransport } from "@modelcontextprotocol/sdk/client/sse.js";
+import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js";
+import { Server } from "@modelcontextprotocol/sdk/server/index.js";
+import { EventSource } from "eventsource";
+import { getRandomPort } from "get-port-please";
+import path from "node:path";
+import { setTimeout as delay } from "node:timers/promises";
+import { describe, expect, it, vi } from "vitest";
+
+import { proxyServer } from "../src/proxyServer.js";
+import { startSSEServer } from "../src/startSSEServer.js";
+
+if (!("EventSource" in global)) {
+  // @ts-expect-error - figure out how to use --experimental-eventsource with vitest
+  global.EventSource = EventSource;
+}
+
+describe("startSSEServer.test.ts", () => {
+  it("proxies messages between SSE and stdio servers", async () => {
+    const stdioTransport = new StdioClientTransport({
+      args: [path.join(__dirname, "./simple-stdio-server.ts")],
+      command: "tsx",
+    });
+
+    const stdioClient = new Client(
+      {
+        name: "mcp-proxy",
+        version: "1.0.0",
+      },
+      {
+        capabilities: {},
+      },
+    );
+
+    await stdioClient.connect(stdioTransport);
+
+    const serverVersion = stdioClient.getServerVersion() as {
+      name: string;
+      version: string;
+    };
+
+    const serverCapabilities = stdioClient.getServerCapabilities() as {
+      capabilities: Record<string, unknown>;
+    };
+
+    const port = await getRandomPort();
+
+    const onConnect = vi.fn();
+    const onClose = vi.fn();
+
+    await startSSEServer({
+      createServer: async () => {
+        const mcpServer = new Server(serverVersion, {
+          capabilities: serverCapabilities,
+        });
+
+        await proxyServer({
+          client: stdioClient,
+          server: mcpServer,
+          serverCapabilities,
+        });
+
+        return mcpServer;
+      },
+      endpoint: "/sse",
+      onClose,
+      onConnect,
+      port,
+    });
+
+    const sseClient = new Client(
+      {
+        name: "sse-client",
+        version: "1.0.0",
+      },
+      {
+        capabilities: {},
+      },
+    );
+
+    const transport = new SSEClientTransport(
+      new URL(`http://localhost:${port}/sse`),
+    );
+
+    await sseClient.connect(transport);
+
+    const result = await sseClient.listResources();
+    expect(result).toEqual({
+      resources: [
+        {
+          name: "Example Resource",
+          uri: "file:///example.txt",
+        },
+      ],
+    });
+
+    expect(
+      await sseClient.readResource({ uri: result.resources[0].uri }, {}),
+    ).toEqual({
+      contents: [
+        {
+          mimeType: "text/plain",
+          text: "This is the content of the example resource.",
+          uri: "file:///example.txt",
+        },
+      ],
+    });
+    expect(await sseClient.subscribeResource({ uri: "xyz" })).toEqual({});
+    expect(await sseClient.unsubscribeResource({ uri: "xyz" })).toEqual({});
+    expect(await sseClient.listResourceTemplates()).toEqual({
+      resourceTemplates: [
+        {
+          description: "Specify the filename to retrieve",
+          name: "Example resource template",
+          uriTemplate: `file://{filename}`,
+        },
+      ],
+    });
+
+    expect(onConnect).toHaveBeenCalled();
+    expect(onClose).not.toHaveBeenCalled();
+
+    await sseClient.close();
+
+    await delay(100);
+
+    expect(onClose).toHaveBeenCalled();
+  });
+});
