+++ test/startStdioServer.test.ts
@@ -6,9 +6,10 @@ import {
 } from "@modelcontextprotocol/sdk/types.js";
 import { EventSource } from "eventsource";
 import { ChildProcess, fork } from "node:child_process";
+import path from "node:path";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import { ServerType } from "./startStdioServer.js";
+import { ServerType } from "../src/startStdioServer.js";
 
 if (!("EventSource" in global)) {
   // @ts-expect-error - figure out how to use --experimental-eventsource with vitest
@@ -43,7 +44,7 @@ describe("startStdioServer.test.ts", () => {
   it("proxies messages between stdio and sse servers", async () => {
     const stdioTransport = new StdioClientTransport({
       args: [
-        "src/simple-stdio-proxy-server.ts",
+        path.join(__dirname, "./simple-stdio-proxy-server.ts"),
         JSON.stringify({
           serverType: ServerType.SSE,
           url: "http://127.0.0.1:3000/sse",
@@ -136,7 +137,7 @@ describe("startStdioServer.test.ts", () => {
   it("proxies messages between stdio and stream able servers", async () => {
     const stdioTransport = new StdioClientTransport({
       args: [
-        "src/simple-stdio-proxy-server.ts",
+        path.join(__dirname, "./simple-stdio-proxy-server.ts"),
         JSON.stringify({
           serverType: ServerType.HTTPStream,
           url: "http://127.0.0.1:3000/mcp",
